// src/app/module/Admin/admin.interface.ts
import { Model, Types } from "mongoose";
import { TBloodGroup, TGender, TUserName } from "../../interface/user";


export type TAdmin = {
    id: string;
    user: Types.ObjectId;
    designation: string;
    name: TUserName;
    gender: TGender;
    dateOfBirth?: Date;
    email: string;
    password: string;
    contactNo: string;
    emergencyContactNo: string;
    bloodGroup?: TBloodGroup;
    presentAddress: string;
    permanentAddress: string;
    profileImg?: string;
    isDeleted: boolean;
};

export interface AdminModel extends Model<TAdmin> {
    isUserExists(id: string): Promise<TAdmin | null>;

}